Arduino sprintf() does not support 64-bit integers

+++ main/Buffer.cpp
@@ -352,7 +352,7 @@ Buffer Buffer::millis_to_hms(int64_t t)
 	int32_t h = t % 24;
 	t -= h;
 	t /= 24;
-	int64_t d = t;
+	int32_t d = t;
 
 	if (d > 0) return BufferImpl::sprintf("%d:%02d:%02d:%02d", d, h, m, s);
 	if (h > 0) return BufferImpl::sprintf("%d:%02d:%02d", h, m, s);
